refactor(extension): isolate direct summary requests
